fix(convex): rename billing-derivation.ts to billingDerivation.ts (hyphens invalid in Convex module paths)
Convex's module-path validator rejects hyphens — only alphanumerics,
underscores, and periods are allowed. Every other convex/lib/*.ts file
already uses camelCase for exactly this reason (projectWindow.ts,
bulkCheckin.ts, etc); this one slipped through kebab-case because none of
tsc/lint/vitest/next-build validate real Convex module paths — only a real
'convex dev'/'convex deploy' push does, which is why this only surfaced in
the E2E harness job. src/lib/billing-derivation.ts (the src-side parity
copy) is unaffected — src/lib uses kebab-case throughout.
